pyproject.toml
[project]
name = "ai-vision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pillow>=10.0.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.0.0",
    "langgraph>=0.0.60",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.20",
    "langsmith>=0.0.80",
    "llama-api-client>=0.1.2",
    "streamlit>=1.46.1",
]
